requirements.txt
tabula-py==2.1.1
tqdm==4.48.2
numpy==1.19.1
pandas==1.1.0